backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-multipart
email-validator
python-dotenv
psycopg2-binary
alembic
Pillow
pydantic-settings
PyJWT
PyPDF2
python-docx
sentence-transformers
faiss-cpu
numpy
scikit-learn
langchain
langchain-community
chromadb
tiktoken
openai
requests
pandas
openpyxl
